*** empty log message ***

+++ magick/segment.c
@@ -1300,7 +1300,7 @@ static void ScaleSpace(const long *histogram,const double tau,
   alpha=1.0/(tau*sqrt((double) (2.0*MagickPI)));
   beta=(-1.0/(2.0*tau*tau));
   for (x=0; x <= (int) DownScale(MaxRGB); x++)
-    gamma[x]=exp((double) (beta*x*x));
+    gamma[x]=exp(beta*x*x);
   for (x=0; x <= (int) DownScale(MaxRGB); x++)
   {
     sum=0.0;
